==> p_07_final_LDA_paper.ipynb <==
## import libraries

#import gensim for topic modelling
import gensim # https://radimrehurek.com/gensim/models/ldamodel.html
from gensim import corpora
from gensim.models import CoherenceModel #to calculate coherence
import time #to have today's date
timestr = time.strftime("%Y%m%d")
import pandas as pd

#get the clean text from the other notebook
%run ./p_03_prepare_text_paper.ipynb
# doc_clean = get_doc_clean() ##all interviews together
#dict_clean_doc_x ##with different lenghts dropped

## drop short documents

var_drop_x = 10 #drop all documents with number of words <= x. Only keep documents >x
doc_clean_x = dict_clean_doc_x['>'+str(var_drop_x)]  #from p_03_prepare text
doc_list_paras_x = dict_doc_list_paras_x['>'+str(var_drop_x)] #from p_03_prepare text

## convert corpus into a matrix

A DT matrix describes the frequency of terms that occur in a collection of documents. In a document-term matrix, rows correspond to documents in the collection and columns correspond to terms

### creating the term dictionary

# Creating the term dictionary of our corpus, where every unique term is assigned an index. 
def get_dictionary(doc_clean):
    dictionary = corpora.Dictionary(doc_clean)
    return(dictionary)

#create dictionaries
dictionary_x = get_dictionary(doc_clean_x)

### creating the document term matrix

# Converting list of documents (corpus) into Document Term Matrix using dictionary prepared above.
#this is also called Bag of Words corpus
def get_doc_term_matrix(dictionary, doc_clean):
    doc_term_matrix = [dictionary.doc2bow(doc) for doc in doc_clean]
    return(doc_term_matrix)

#get document term matrix
doc_term_matrix_x = get_doc_term_matrix(dictionary_x, doc_clean_x)

## Running the LDA Model

create an object for LDA model and train it on Document-Term matrix

# Creating the object for LDA model using gensim library + # Running and Trainign LDA model on the document term matrix.
def get_ldamodel(var_alpha, var_no_topics, var_passes, doc_term_matrix, dictionary):
    ldamodel = gensim.models.ldamodel.LdaModel(doc_term_matrix, num_topics=var_no_topics, alpha=var_alpha,
                id2word = dictionary, random_state=1, passes=var_passes,minimum_probability=0) 
    #use random_state for reproducity
    return(ldamodel)

#choose parameters, run model:
var_alpha = 0.01
var_no_topics = 22
var_passes = 50 # default: passes=1
ldamodel = get_ldamodel(var_alpha, var_no_topics, var_passes, doc_term_matrix_x, dictionary_x)

#compare model results with the ones from the notebook p_05_set_parameters_LDA <br>
coherence_model_lda = CoherenceModel(model=ldamodel, \
                texts=doc_clean_x, dictionary=dictionary_x, coherence='c_v')
coherence_lda = coherence_model_lda.get_coherence() #get the coherence score
print(coherence_lda)

#do list with the topics keywords for graphs
list_topics_keywords = list()
for x in range(0, var_no_topics):
    wp = ldamodel.show_topic(x, topn=20) #default: topn=10, number of words
    topic_keywords = ", ".join([word for word, prop in wp])
    list_topics_keywords.append(topic_keywords)

#make dataframe with original topic numbers and the keywords
df_ori_num_keyw = pd.DataFrame({'ori_topic_num':range(0, len(list_topics_keywords)), 'keywords':list_topics_keywords})
df_ori_num_keyw.to_csv('.//results//topics_group//topic_original_numbers_keywords_'+timestr+'_V01.csv',
    sep=';', decimal=',')

list_topics_keywords
