==> p_13_do_master_df_paper.ipynb <==
import pandas as pd
import time #to have today's date
timestr = time.strftime("%Y%m%d")

#get the LDA model from another notebook
%run ./p_07_final_LDA_paper.ipynb

#doc_term_matrix_x 
#ldamodel =
#var_drop_x #drop all documents with number of words <= x. Only keep documents >x
#doc_clean_x 
#doc_list_paras_x

## do dataframe with all the topics+categories+keywords and the text+questions per document

### create the dataframe

#create dataframe with empty columns
topics_per_doc_df = pd.DataFrame()
list_topics_columns = list() #to use later
for x in range(0,var_no_topics):
    topics_per_doc_df["perc_cont_"+str(x)] = []
    topics_per_doc_df["keywords_"+str(x)] = []
    list_topics_columns.append("perc_cont_"+str(x))

# Get relevant topics for each document, this takes a while
for i, row in enumerate(ldamodel[doc_term_matrix_x]): #this provides a list with all topics and their percentage contribution
    # Get all the topics, Perc Contribution and Keywords for each document
    for j in row:
        topic_num = j[0]
        prop_topic = j[1]
        wp = ldamodel.show_topic(topic_num)
        topic_keywords = ", ".join([word for word, prop in wp])
        topics_per_doc_df.loc[i, "perc_cont_"+str(topic_num)] = prop_topic
        topics_per_doc_df.loc[i, "keywords_"+str(topic_num)] = topic_keywords

# Add original text to the end of the output
contents = pd.Series(doc_list_paras_x)
topics_per_doc_df = pd.concat([topics_per_doc_df, contents], axis=1)
topics_per_doc_df.rename(columns={0:'Text'}, inplace=True) #rename the added column with the original text

#no need anymore, question is already in text
#if available, also add the question that was asked to each document
#df_questions_replies = pd.read_csv('.//results//topics_group//questions_replies_'+timestr+'_V01.csv',
#    sep=';', decimal=',', index_col=0)
df_docs_topics = topics_per_doc_df#.merge(df_questions_replies)

#merge to get information on country, class etc. for each document
df_docs_topics_country_class = df_all_interviews.merge(df_docs_topics) #do I need df_all_interviews_x? -->no!
#P: the text "I don't have information on that." is said two times.
#thus, when we merge the dataframe, it appears four times. 
#S: drop duplicate rows:
df_docs_topics_country_class = df_docs_topics_country_class.drop_duplicates().\
        reset_index(drop=True)

df_docs_topics_country_class.to_csv('.//results//master_dfs//docs_oldNum_topics_country_class_keywords_text_q_'+timestr+'_V01.csv',
    sep=';', decimal=',')

## change numbers to new topic numbers

#get dataframe with the categories (numbers & names), topics (numbers (ori&new) & names), and keywords and lambda
df_topics_categories = pd.read_csv\
('.//results//master_dfs//category_numname_topics_numname_orinew_keywords_lambda_'+timestr+'_V01.csv',
    sep=';', decimal=',', index_col=0)

df_topic_nums = df_topics_categories[['ori_topic_num', 'Topic_num']].sort_values('ori_topic_num')
#the keys are the old (original) topic numbers, the value are the new ones.
dict_t_nums = dict(zip(df_topic_nums.ori_topic_num, df_topic_nums.Topic_num))
#so for all the keys (for x in dict_t_nums), if x is the number, change x to dict_t_nums[x]

#change the numbers
df_docs_newNum_topics_country_class = df_docs_topics_country_class.copy()

for x in df_docs_newNum_topics_country_class.columns:
    list_x_split = x.split('_')
    if list_x_split[0] == 'perc': #so column is 'perc_cont_x'
        var_number = int(list_x_split[2])
        df_docs_newNum_topics_country_class.rename(columns = {x:'t_'+str(dict_t_nums[var_number])+'_perc_cont'}, inplace = True)
    if list_x_split[0] == 'keywords': #so column is 'perc_cont_x'
        var_number = int(list_x_split[1])
        df_docs_newNum_topics_country_class.rename(columns = {x:'t_'+str(dict_t_nums[var_number])+'_keywords'}, inplace = True)

#rearrange column names:
list_perc = []
for x in range(1, df_topic_nums['ori_topic_num'].max()+2):
    list_perc.append('t_'+str(x)+'_perc_cont')
    list_perc.append('t_'+str(x)+'_keywords')

list_columns = ['Country', 'no_interview', 'class', 'Text'] + list_perc #+ ['Question']

df_docs_newNum_topics_country_class_order = df_docs_newNum_topics_country_class[list_columns]

df_docs_newNum_topics_country_class_order.tail(3)

## export dataframe

df_docs_newNum_topics_country_class_order.to_csv('.//results//master_dfs//docs_newNum_topics_country_class_keywords_text_q_'+timestr+'_V01.csv',
    sep=';', decimal=',')
