==> p_15_do_dfs_country+class_category_paper.ipynb <==
import pandas as pd
import time #to have today's date
timestr = time.strftime("%Y%m%d")
import numpy as np

#import csv. 
df_topics_categories = pd.read_csv\
('.//results//master_dfs//category_numname_topics_numname_orinew_keywords_lambda_'+timestr+'_V01.csv',
    sep=';', decimal=',', index_col=0)

# convert topic numbers per category into a list
groups = df_topics_categories.groupby('Category_num')['Topic_num'].apply(list)
#make dictionary
dict_t_c_num = dict(groups)

#merge the topics per group
def topics_into_categories(df_t_c):
    for x in dict_t_c_num: #all the category numbers
        df_t_c['c_'+str(x)] = 0
        for y in dict_t_c_num[x]:
            df_t_c['c_'+str(x)] += df_t_c[str(y)]
            df_t_c.drop(columns=str(y), inplace=True)
        df_t_c.rename(columns={'c_'+str(x):x}, inplace=True)
    return(df_t_c)

## do dataframes with categories per country and actor group

### per country

df_country_topics_share = pd.read_csv('.//results//master_dfs//country_topics_share_'+timestr+'_V01.csv',
    sep=';', decimal=',', index_col=0)

df_country_categories_share = topics_into_categories(df_country_topics_share.copy())

df_country_categories_share.to_csv('.//results//master_dfs//country_categories_share_'+timestr+'_V01.csv',
    sep=';', decimal=',')

### per actor group

df_class_topics_share = pd.read_csv('.//results//master_dfs//class_topics_share_'+timestr+'_V01.csv',
    sep=';', decimal=',', index_col=0)

df_class_categories_share = topics_into_categories(df_class_topics_share.copy())

df_class_categories_share.to_csv('.//results//master_dfs//class_categories_share_'+timestr+'_V01.csv',
    sep=';', decimal=',')

### per country AND actor group

df_country_class_topics_share = pd.read_csv('.//results//master_dfs//country_AND_class_topics_share_'+timestr+'_V01.csv',
    sep=';', decimal=',', index_col=0)

df_country_class_categories_share = topics_into_categories(df_country_class_topics_share.copy())

df_country_class_categories_share.to_csv('.//results//master_dfs//country_AND_class_categories_share_'+timestr+'_V01.csv',
    sep=';', decimal=',')

## for shares between countries for each topic, so each topic is 100%

df_country_topics_share_T = pd.read_csv('.//results//master_dfs//country_topics_share_T_'+timestr+'_V01.csv',
    sep=';', decimal=',', index_col=0)
df_country_categories_share_T = topics_into_categories(df_country_topics_share_T.copy())
df_country_categories_share_T.to_csv('.//results//master_dfs//country_categories_share_T_'+timestr+'_V01.csv',
    sep=';', decimal=',')

df_class_topics_share_T = pd.read_csv('.//results//master_dfs//class_topics_share_T_'+timestr+'_V01.csv',
    sep=';', decimal=',', index_col=0)
df_class_categories_share_T = topics_into_categories(df_class_topics_share_T.copy())
df_class_categories_share_T.to_csv('.//results//master_dfs//class_categories_share_T_'+timestr+'_V01.csv',
    sep=';', decimal=',')

df_country_class_topics_share_T = pd.read_csv('.//results//master_dfs//country_AND_class_topics_share_T_'+timestr+'_V01.csv',
    sep=';', decimal=',', index_col=0)
df_country_class_categories_share_T = topics_into_categories(df_country_class_topics_share_T.copy())
df_country_class_categories_share_T.to_csv('.//results//master_dfs//country_AND_class_categories_share_T_'+timestr+'_V01.csv',
    sep=';', decimal=',')
