==> p_20_table_doc_per_topic_per_country_paper.ipynb <==
import pandas as pd
import time #to have today's date
timestr = time.strftime("%Y%m%d")

# update to newest version of the dataframes!
# for df_topics_categories:
var_time_d = '20220517'

df_docs_newNum_topics_country_class_order = \
    pd.read_csv('.//results//master_dfs//docs_newNum_topics_country_class_keywords_text_q_'+var_time_d+'_V01.csv',
    sep=';', decimal=',', index_col=0)

list_country_oder = ['DE', 'UK', 'CL', 'CO', 'US', 'KE', 'ZA', 'ID', 'IN', 'VN', 'PH', 'PK']
list_num = list(range(1, 23))

list_num

list_columns = list_country_oder.copy()
list_columns.insert(0, 't_num')

list_columns

df_doc_per_topic_per_country = pd.DataFrame() #the final dataset
df_docs = df_docs_newNum_topics_country_class_order.copy()
m = 0
for x in list_country_oder:
    df_docs_one_country = pd.DataFrame(columns=['t_num', x+'_Text'])
    df_c = df_docs[df_docs['Country']==x]
    for n in list_num:
        var_t = 't_'+str(n)+'_perc_cont'
        df_t = df_c[['Text', var_t]]
        df_t = df_t.sort_values(by=var_t, ascending=False) #sort by contribution
        df_t.rename(columns={var_t:x+'_contribution', 'Text':x+'_Text'}, inplace=True) #add the country to the column names
        df_t_short = df_t.head(5) #only show top 5 documents
        list_k = []
        for k in list(range(1,6)):
            list_k.append('#'+str(n)+'.'+str(k))
        df_t_short['t_num'] = list_k #add column to perform merge on later on
        df_docs_one_country = df_docs_one_country.append(df_t_short)
        df_docs_one_country.reset_index(drop=True, inplace=True)
    if m == 0: #for the first, take as dataframe
        df_doc_per_topic_per_country = df_docs_one_country.copy()
    else: #later merge with the existing dataframe on the t_num columns
        df_doc_per_topic_per_country = pd.merge(df_doc_per_topic_per_country, df_docs_one_country, on ='t_num')
    m += 1

df_doc_per_topic_per_country.to_csv('.//results//table_docs_per_topic_per_country//table_docs_per_topic_per_country_'+timestr+'_V01.csv',
    sep=';', decimal=',')
