==> p_22_overview_of_interviews_qa_short_dropped_paper.ipynb <==
## Import Libraries

import pandas as pd #to work with DataFrames

import time #to have today's date
timestr = time.strftime("%Y%m%d")

var_time_a = '20220517'
df_master_docs = pd.read_csv('.//results//master_dfs//docs_newNum_topics_country_class_keywords_text_q_'+var_time_a+'_V01.csv',
    sep=';', decimal=',', index_col=0)

## for countries

list_new_country_order = ['DE', 'UK', 'CL', 'CO', 'US', 'KE', 'ZA', 'ID', 'IN', 'VN', 'PH', 'PK']

df_overview_10_countries = pd.DataFrame()

# add countries and number of interviews to df:
for x in list_new_country_order:
    df_x = df_master_docs[df_master_docs['Country']==x]
    var_no_interviews = df_x['no_interview'].max()
    var_no_docs = len(df_x)
    var_no_words = 0
    for t in df_x['Text']:
        tr = t.replace('  ', '').replace('<q/start>', '').replace('<q/end>', '')
        var_no_words += len(tr.split(' '))
    #calculate mean values
    var_docs_int = round(var_no_docs/var_no_interviews, 1)
    var_words_int = round(var_no_words/var_no_interviews, 1)
    var_words_doc = round(var_no_words/var_no_docs , 1)
    #add row to df
    df_overview_10_countries = df_overview_10_countries.append({'Country':x, \
            'Interviews':var_no_interviews, 'Documents':var_no_docs, 'Words':var_no_words, \
            'docs_per_int':var_docs_int, 'words_per_int':var_words_int, 'words_per_doc':var_words_doc}, ignore_index=True)

#add total to end of df
var_total_int = df_overview_10_countries['Interviews'].sum()
var_total_docs = df_overview_10_countries['Documents'].sum()
var_total_words = df_overview_10_countries['Words'].sum()
var_total_docs_int = round(var_total_docs/var_total_int, 1)
var_total_words_int = round(var_total_words/var_total_int, 1)
var_total_words_doc = round(var_total_words/var_total_docs, 1)

df_overview_10_countries = df_overview_10_countries.append({'Country':'TOTAL', \
            'Interviews':var_total_int, 'Documents':var_total_docs, 'Words':var_total_words, \
            'docs_per_int':var_total_docs_int, 'words_per_int':var_total_words_int, \
            'words_per_doc':var_total_words_doc}, ignore_index=True)

df_overview_10_countries.to_csv('.\\results\overview_interviews\qa_long_overview_interviews_by_country_'+timestr+'.csv',
        sep=';', decimal=',')

## for actor groups

list_order_actors = ['econ', 'poli', 'socint', 'socnat']

df_overview_10_actors = pd.DataFrame()

for x in list_order_actors:
    df_x = df_master_docs[df_master_docs['class']==x]
    df_x_r = df_x.reset_index(drop=True)
    df_x_gr = df_x_r.reset_index(drop=True).groupby(['Country', 'no_interview'])[['Text']].sum()
    var_no_interviews = len(df_x_gr)
    var_no_docs = len(df_x)
    var_no_words = 0
    for t in df_x['Text']:
        tr = t.replace('  ', '').replace('<q/start>', '').replace('<q/end>', '')
        var_no_words += len(tr.split(' '))
    #calculate mean values
    var_docs_int = round(var_no_docs/var_no_interviews, 1)
    var_words_int = round(var_no_words/var_no_interviews, 1)
    var_words_doc = round(var_no_words/var_no_docs , 1)
    #add row to df
    df_overview_10_actors = df_overview_10_actors.append({'Class':x, \
            'Interviews':var_no_interviews, 'Documents':var_no_docs, 'Words':var_no_words, \
            'docs_per_int':var_docs_int, 'words_per_int':var_words_int, 'words_per_doc':var_words_doc}, ignore_index=True)

#add total to end of df
var_total_int = df_overview_10_actors['Interviews'].sum()
var_total_docs = df_overview_10_actors['Documents'].sum()
var_total_words = df_overview_10_actors['Words'].sum()
var_total_docs_int = round(var_total_docs/var_total_int, 1)
var_total_words_int = round(var_total_words/var_total_int, 1)
var_total_words_doc = round(var_total_words/var_total_docs, 1)

df_overview_10_actors = df_overview_10_actors.append({'Class':'TOTAL', \
            'Interviews':var_total_int, 'Documents':var_total_docs, 'Words':var_total_words, \
            'docs_per_int':var_total_docs_int, 'words_per_int':var_total_words_int, \
            'words_per_doc':var_total_words_doc}, ignore_index=True)

df_overview_10_actors

df_overview_10_actors.to_csv('.\\results\overview_interviews\qa_long_overview_interviews_by_classification_'+timestr+'.csv',
    sep=';', decimal=',')
